refactor: changes edit profile button color
