+++ sharetunes-app/src/components/common/buttons/PrimaryButton/PrimaryButton.tsx
@@ -3,12 +3,16 @@ import React from 'react';
 import classes from './primaryButton.module.scss';
 
 interface Props {
-    text: String,
-    onButtonClick: () => void,
-    buttonColor?: 'black' | 'blue' | 'green'
+    text: String;
+    onButtonClick: () => void;
+    buttonColor?: 'black' | 'blue' | 'green' | 'editProfileBtn';
 }
 
-const PrimaryButton : React.FC<Props> = ({text, onButtonClick, buttonColor}) => {
+const PrimaryButton: React.FC<Props> = ({
+    text,
+    onButtonClick,
+    buttonColor,
+}) => {
     let buttonStyle;
 
     if (buttonColor === undefined) {
@@ -17,16 +21,18 @@ const PrimaryButton : React.FC<Props> = ({text, onButtonClick, buttonColor}) => 
         buttonStyle = classes.blackStyle;
     } else if (buttonColor === 'green') {
         buttonStyle = classes.greenStyle;
+    } else if (buttonColor === 'editProfileBtn') {
+        buttonStyle = classes.editProfileBtn;
     }
-    
+
     return (
-        <div 
-            onClick = {onButtonClick}
-            className={[buttonStyle, classes.PrimaryButton].join(" ")}
+        <div
+            onClick={onButtonClick}
+            className={[buttonStyle, classes.PrimaryButton].join(' ')}
         >
             {text}
         </div>
-    )
-}
+    );
+};
 
 export default PrimaryButton;
