+++ sharetunes-app/src/components/common/buttons/PrimaryButton/primaryButton.module.scss
@@ -37,3 +37,11 @@
 .greenStyle:hover {
     background-color: #128139;
 }
+
+.editProfileBtn{
+    border: 2px solid #c3c3c3 
+}
+
+.editProfileBtn:hover{
+    background-color: #a8a8a8
+}
