+++ sharetunes-app/src/components/pages/ProfilePage/profileView.module.scss
@@ -5,8 +5,8 @@
 }
 .ProfileImg {
     border-radius: 50%;
-    width: 300px;
-    height: 300px;
+    width: 200px;
+    height: 200px;
     margin-left:80px;
     margin-right:80px;
 }
